Migrate discovery execution into discovery module
