+++ sqrl-engines/sqrl-engine-core/src/main/java/com/datasqrl/engine/database/DatabaseEngineConfiguration.java
@@ -4,7 +4,6 @@
 package com.datasqrl.engine.database;
 
 import com.datasqrl.error.ErrorCollector;
-import com.datasqrl.metadata.MetadataStoreProvider;
 import com.datasqrl.engine.EngineConfiguration;
 import com.datasqrl.engine.ExecutionEngine;
 import lombok.NonNull;
@@ -14,7 +13,6 @@ import lombok.NonNull;
  */
 public interface DatabaseEngineConfiguration extends EngineConfiguration {
 
-  MetadataStoreProvider getMetadataStore();
 
   @Override
   DatabaseEngine initialize(@NonNull ErrorCollector errors);
