+++ sqrl-engines/sqrl-engine-core/src/main/java/com/datasqrl/engine/database/DatabaseEngineFactory.java
@@ -3,13 +3,9 @@ package com.datasqrl.engine.database;
 import com.datasqrl.config.SqrlConfig;
 import com.datasqrl.engine.EngineFactory;
 import com.datasqrl.engine.ExecutionEngine;
-import com.datasqrl.error.ErrorCollector;
-import com.datasqrl.metadata.MetadataStoreProvider;
 import lombok.NonNull;
 
 public interface DatabaseEngineFactory extends EngineFactory {
-
-  MetadataStoreProvider getMetadataStore(@NonNull SqrlConfig config);
 
   @Override
   DatabaseEngine initialize(@NonNull SqrlConfig config);
@@ -17,7 +13,4 @@ public interface DatabaseEngineFactory extends EngineFactory {
   default ExecutionEngine.Type getEngineType() {
     return ExecutionEngine.Type.DATABASE;
   }
-
-
-
 }
