+++ sqrl-engines/sqrl-engine-core/src/main/java/com/datasqrl/engine/stream/StreamEngine.java
@@ -4,19 +4,8 @@
 package com.datasqrl.engine.stream;
 
 import com.datasqrl.engine.ExecutionEngine;
-import com.datasqrl.engine.stream.monitor.DataMonitor;
 import java.io.Closeable;
 
 public interface StreamEngine extends Closeable, ExecutionEngine {
-
-  /**
-   * This method must be implemented if the engine supports {@link com.datasqrl.engine.EngineCapability#DATA_MONITORING}
-   * otherwise it can be ignored.
-   *
-   * @return
-   */
-  default DataMonitor createDataMonitor() {
-    throw new UnsupportedOperationException("Capability not supported by engine");
-  }
 
 }
