+++ sqrl-engines/sqrl-engine-core/src/main/java/com/datasqrl/engine/stream/inmemory/InMemoryStreamConfiguration.java
@@ -6,6 +6,7 @@ package com.datasqrl.engine.stream.inmemory;
 import com.datasqrl.error.ErrorCollector;
 import com.datasqrl.engine.EngineConfiguration;
 import com.datasqrl.engine.ExecutionEngine;
+import com.google.auto.service.AutoService;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -17,6 +18,7 @@ import lombok.NonNull;
 @Builder
 @Getter
 @NoArgsConstructor
+@AutoService(EngineConfiguration.class)
 public class InMemoryStreamConfiguration implements EngineConfiguration {
 
   public static final String ENGINE_NAME = "memStream";
