--- sqrl-engines/sqrl-engine-core/src/main/resources/META-INF/services/com.datasqrl.engine.EngineConfiguration
@@ -1,2 +1,0 @@
-com.datasqrl.engine.stream.inmemory.InMemoryStreamConfiguration
-com.datasqrl.engine.database.inmemory.InMemoryDatabaseConfiguration
