+++ sqrl-engines/sqrl-engine-flink/src/main/java/com/datasqrl/engine/stream/flink/FlinkEngineFactory.java
@@ -26,12 +26,7 @@ public class FlinkEngineFactory implements EngineFactory {
 
   @Override
   public AbstractFlinkStreamEngine initialize(@NonNull SqrlConfig config) {
-    return new LocalFlinkStreamEngineImpl(new ExecutionEnvironmentFactory(getFlinkConfiguration(config)),
-        config);
-  }
-
-  public Map<String,String> getFlinkConfiguration(@NonNull SqrlConfig config) {
-    return SqrlConfigUtil.toStringMap(config, EngineFactory.getReservedKeys());
+    return new LocalFlinkStreamEngineImpl(config);
   }
 
 }
