+++ sqrl-engines/sqrl-engine-flink/src/main/java/com/datasqrl/engine/stream/flink/LocalFlinkStreamEngineImpl.java
@@ -9,13 +9,7 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 public class LocalFlinkStreamEngineImpl extends AbstractFlinkStreamEngine {
 
-  public LocalFlinkStreamEngineImpl(ExecutionEnvironmentFactory execFactory,
-      SqrlConfig config) {
-    super(execFactory, config);
-  }
-
-  public FlinkStreamBuilder createJob() {
-    return new FlinkStreamBuilder(this,
-        execFactory.createEnvironment());
+  public LocalFlinkStreamEngineImpl(SqrlConfig config) {
+    super(config);
   }
 }
