+++ sqrl-engines/sqrl-engine-jdbc/src/main/java/com/datasqrl/engine/database/relational/JDBCEngineFactory.java
@@ -1,19 +1,12 @@
 package com.datasqrl.engine.database.relational;
 
 import com.datasqrl.config.SqrlConfig;
-import com.datasqrl.config.serializer.KryoProvider;
-import com.datasqrl.config.serializer.SerializerProvider;
 import com.datasqrl.engine.EngineFactory;
-import com.datasqrl.engine.database.DatabaseEngine;
 import com.datasqrl.engine.database.DatabaseEngineFactory;
-import com.datasqrl.engine.database.relational.metadata.JDBCMetadataStore;
 import com.datasqrl.io.impl.jdbc.JdbcDataSystemConnector;
 import com.datasqrl.io.impl.jdbc.JdbcDataSystemConnectorFactory;
-import com.datasqrl.metadata.MetadataStore;
-import com.datasqrl.metadata.MetadataStoreProvider;
 import com.google.auto.service.AutoService;
 import lombok.NonNull;
-import lombok.Value;
 
 @AutoService(EngineFactory.class)
 public class JDBCEngineFactory implements DatabaseEngineFactory {
@@ -25,11 +18,6 @@ public class JDBCEngineFactory implements DatabaseEngineFactory {
   }
 
   @Override
-  public MetadataStoreProvider getMetadataStore(@NonNull SqrlConfig config) {
-    return new StoreProvider(getConnector(config));
-  }
-
-  @Override
   public JDBCEngine initialize(@NonNull SqrlConfig config) {
     return new JDBCEngine(getConnector(config));
   }
@@ -37,18 +25,4 @@ public class JDBCEngineFactory implements DatabaseEngineFactory {
   private JdbcDataSystemConnector getConnector(@NonNull SqrlConfig config) {
     return new JdbcDataSystemConnectorFactory().getConnector(config);
   }
-
-  @Value
-  public static class StoreProvider implements MetadataStoreProvider {
-
-    JdbcDataSystemConnector connection;
-    SerializerProvider serializer = new KryoProvider(); //TODO: make configurable
-
-    @Override
-    public MetadataStore openStore() {
-      return new JDBCMetadataStore(connection, serializer.getSerializer());
-    }
-
-  }
-
 }
