+++ sqrl-io/sqrl-io-jdbc/pom.xml
@@ -21,9 +21,5 @@
       <artifactId>sqrl-io-core</artifactId>
     </dependency>
 
-    <dependency>
-      <groupId>com.esotericsoftware.kryo</groupId>
-      <artifactId>kryo</artifactId>
-    </dependency>
   </dependencies>
 </project>
