+++ sqrl-tools/sqrl-discovery/src/main/java/com/datasqrl/discovery/DataDiscoveryFactory.java
@@ -1,12 +1,18 @@
 package com.datasqrl.discovery;
 
-import com.datasqrl.config.EngineSettings;
+import static com.datasqrl.config.PipelineFactory.ENGINES_PROPERTY;
+
 import com.datasqrl.config.PipelineFactory;
 import com.datasqrl.config.SqrlConfig;
-import com.datasqrl.engine.stream.StreamEngine;
+import com.datasqrl.engine.EngineFactory;
+import com.datasqrl.engine.ExecutionEngine.Type;
+import com.datasqrl.engine.database.DatabaseEngineFactory;
+import com.datasqrl.engine.database.inmemory.InMemoryDatabaseFactory;
+import com.datasqrl.engine.database.relational.JDBCEngineFactory;
 import com.datasqrl.error.ErrorCollector;
-import java.nio.file.Path;
-import java.util.List;
+import com.datasqrl.metadata.InMemoryMetadataStore;
+import com.datasqrl.metadata.JdbcMetadataStore;
+import com.datasqrl.metadata.MetadataStoreProvider;
 import java.util.Optional;
 import lombok.NonNull;
 import org.apache.commons.lang3.StringUtils;
@@ -23,13 +29,36 @@ public class DataDiscoveryFactory {
     Optional<String> metadataEngine = discoveryConfig.asString(DATABASE_ENGINE_KEY).validate(
         StringUtils::isNotBlank,"Cannot be empty").getOptional();
     DataDiscovery discovery = new DataDiscovery(errors, pipelineFactory.getStreamEngine(),
-        pipelineFactory.getMetaDataStoreProvider(metadataEngine));
+        getMetaDataStoreProvider(config.getSubConfig(ENGINES_PROPERTY), metadataEngine),
+        pipelineFactory.getConfig());
     return discovery;
   }
 
   public static DataDiscovery fromPipeline(@NonNull PipelineFactory pipelineFactory, ErrorCollector errors) {
     return new DataDiscovery(errors, pipelineFactory.getStreamEngine(),
-        pipelineFactory.getMetaDataStoreProvider(Optional.empty()));
+        getMetaDataStoreProvider(pipelineFactory.getConfig(), Optional.empty()),
+        pipelineFactory.getConfig());
   }
 
+  public static MetadataStoreProvider getMetaDataStoreProvider(SqrlConfig baseEngineConfig, Optional<String> engineIdentifier) {
+    for (String engineId : baseEngineConfig.getKeys()) {
+      SqrlConfig engineConfig = baseEngineConfig.getSubConfig(engineId);
+      EngineFactory engineFactory = EngineFactory.fromConfig(engineConfig);
+      if (engineIdentifier.map(id -> id.equalsIgnoreCase(engineId))
+          .orElse(engineFactory.getEngineType()== Type.DATABASE)) {
+        baseEngineConfig.getErrorCollector().checkFatal(engineFactory instanceof DatabaseEngineFactory,
+            "Selected or default engine [%s] for metadata is not a database engine", engineId);
+        if (engineFactory instanceof JDBCEngineFactory) {
+          JdbcMetadataStore metadataStore = new JdbcMetadataStore();
+          return metadataStore.getMetadataStore(engineConfig);
+        } else if (engineFactory instanceof InMemoryDatabaseFactory) {
+          InMemoryMetadataStore inMemoryMetadataStore = new InMemoryMetadataStore();
+          return inMemoryMetadataStore.getMetadataStore(engineConfig);
+        } else {
+          throw new RuntimeException("Unknown engine type for mata data store");
+        }
+      }
+    }
+    throw baseEngineConfig.getErrorCollector().exception("Could not find database engine for metadata");
+  }
 }
