+++ sqrl-tools/sqrl-discovery/src/main/java/com/datasqrl/engine/database/inmemory/InMemoryDatabaseConfiguration.java
@@ -3,13 +3,14 @@
  */
 package com.datasqrl.engine.database.inmemory;
 
+import com.datasqrl.engine.EngineConfiguration;
 import com.datasqrl.error.ErrorCollector;
-import com.datasqrl.metadata.MetadataStore;
-import com.datasqrl.metadata.MetadataStoreProvider;
 import com.datasqrl.engine.database.DatabaseEngine;
 import com.datasqrl.engine.database.DatabaseEngineConfiguration;
+import com.google.auto.service.AutoService;
 import lombok.NonNull;
 
+@AutoService(EngineConfiguration.class)
 public class InMemoryDatabaseConfiguration implements DatabaseEngineConfiguration {
 
   public static final String ENGINE_NAME = "hashmap";
@@ -23,18 +24,4 @@ public class InMemoryDatabaseConfiguration implements DatabaseEngineConfiguratio
   public DatabaseEngine initialize(@NonNull ErrorCollector errors) {
     return new InMemoryDatabase();
   }
-
-  @Override
-  public MetadataStoreProvider getMetadataStore() {
-    return new StoreProvider();
-  }
-
-  public static class StoreProvider implements MetadataStoreProvider {
-
-    @Override
-    public MetadataStore openStore() {
-      return new InMemoryMetadataStore();
-    }
-  }
-
 }
