+++ sqrl-tools/sqrl-discovery/src/main/java/com/datasqrl/engine/database/inmemory/InMemoryDatabaseFactory.java
@@ -4,10 +4,6 @@ import com.datasqrl.config.SqrlConfig;
 import com.datasqrl.engine.EngineFactory;
 import com.datasqrl.engine.database.DatabaseEngine;
 import com.datasqrl.engine.database.DatabaseEngineFactory;
-import com.datasqrl.engine.database.inmemory.InMemoryDatabaseConfiguration.StoreProvider;
-import com.datasqrl.error.ErrorCollector;
-import com.datasqrl.metadata.MetadataStore;
-import com.datasqrl.metadata.MetadataStoreProvider;
 import com.google.auto.service.AutoService;
 import lombok.NonNull;
 
@@ -25,18 +21,4 @@ public class InMemoryDatabaseFactory implements DatabaseEngineFactory {
   public DatabaseEngine initialize(@NonNull SqrlConfig config) {
     return new InMemoryDatabase();
   }
-
-  @Override
-  public MetadataStoreProvider getMetadataStore(@NonNull SqrlConfig config) {
-    return new InMemoryDatabaseConfiguration.StoreProvider();
-  }
-
-  public static class StoreProvider implements MetadataStoreProvider {
-
-    @Override
-    public MetadataStore openStore() {
-      return new InMemoryMetadataStore();
-    }
-  }
-
 }
