+++ sqrl-tools/sqrl-discovery/src/main/java/com/datasqrl/io/util/StreamInputPreparerImpl.java
