+++ sqrl-tools/sqrl-discovery/src/main/java/com/datasqrl/metadata/MetadataEngine.java
@@ -1,0 +1,8 @@
+package com.datasqrl.metadata;
+
+import com.datasqrl.config.SqrlConfig;
+import lombok.NonNull;
+
+public interface MetadataEngine {
+  MetadataStoreProvider getMetadataStore(@NonNull SqrlConfig config);
+}
