+++ sqrl-tools/sqrl-discovery/src/main/java/com/datasqrl/util/BufferedLatestSelector.java
